Turned SQL sync on and changed data types.

+++ models/RaspberryPi.js
@@ -2,7 +2,7 @@
 
 module.exports = function(sql, Sql) {
 var RaspberryPi = sql.define("RaspberryPi", {
-		serialNumber: {type: Sql.INTEGER, primaryKey: true},
+		serialNumber: {type: Sql.BIGINT, primaryKey: true},
         machineID: {type: Sql.INTEGER},
         machineType: {type: Sql.STRING}
 }, {timestamps: false, freezeTableName: true});
